jenkins test 11

+++ properties/jenkins_bulid.sh
@@ -1,3 +1,9 @@
+#! /bin/bash
+
 docker build -t develop .
-docker rmi -f $(docker images -q -f dangling=true)
+if [ $(docker images -q -f dangling=true) > 0 ]; then
+    docker rmi -f $(docker images -q -f dangling=true)
+else
+  echo "There is no dangling images"
+fi
 docker run -p 8180:8180 -e server_mode=develop develop
